feat: Implement season creation and default league settings

+++ src/main/kotlin/com/pokerleaguebackend/model/PlacePoint.kt
@@ -1,7 +1,15 @@
 package com.pokerleaguebackend.model
 
 import com.fasterxml.jackson.annotation.JsonBackReference
-import jakarta.persistence.*
+import jakarta.persistence.Entity
+import jakarta.persistence.FetchType
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.JoinColumn
+import jakarta.persistence.ManyToOne
+import jakarta.persistence.Table
+import java.math.BigDecimal
 
 @Entity
 @Table(name = "place_points")
@@ -11,7 +19,7 @@ data class PlacePoint(
     val id: Long = 0,
 
     val place: Int,
-    val points: Int,
+    val points: BigDecimal,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "league_settings_id")
